+++ src/V1/guards/index.ts
@@ -32,4 +32,4 @@
 // POSSIBILITY OF SUCH DAMAGE.
 //
 
-export * from "./isUuid";
+export * from "./isUuidType";