+++ src/V1/guards/isUuidType.ts
@@ -31,22 +31,19 @@
 // ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 //
-import { isUuid, Uuid } from "../";
+import { Uuid } from "..";
 
-describe("isUuid()", () => {
-    it("accepts a Uuid type", () => {
-        const inputValue = new Uuid("123e4567-e89b-12d3-a456-426655440000");
-        expect(isUuid(inputValue)).toBeTrue();
-    });
+/**
+ * A type-guard to make sure that you're dealing with a type-safe Uuid
+ */
+export function isUuidType(input: any): input is Uuid {
+    if (typeof(input) !== "object") {
+        return false;
+    }
 
-    it("rejects other objects", () => {
-        const inputValue = {};
-        expect(isUuid(inputValue)).toBeFalse();
+    if (input.hex === undefined) {
+        return false;
+    }
 
-    });
-
-    it("rejects UUID strings", () => {
-        const inputValue = "123e4567-e89b-12d3-a456-426655440000";
-        expect(isUuid(inputValue)).toBeFalse();
-    });
-});
+    return true;
+}